Change usage paradigm on creating invoices
create_free_form_invoice and create_invoice_based_on_tracked_time_and_expenses now take dataclass objects and we translate them to dicts on behalf of the user. This is because there is minimal difference between the two eventual POST requests and the dataclass is the best way to enforce field names and required fields.

Co-Authored-By: Bradley van Ree <47906258+bradbase@users.noreply.github.com>